Initialize tatami representations as external pointers from R objects. (#21)
This represents a paradigm shift for the next beachmat release. We rely on
beachmat to construct an external pointer to a tatami matrix, which then gets
passed to third-party packages that are compiled against the tatami interface.
In this manner, the concrete tatami representations are only compiled once in
the Bioc ecosystem; each third-party package only uses the interface.

The construction of the external pointer uses initialization functions that 
were initially in the LTLA/arbalist package. This uses S4 dispatch to call
the appropriate C++ function based on the type of R object. We don't use the
existing raticate libraries to handle parsing, as it is often easier and more
reliable to do the work in R (e.g., to figure out the generics being used 
in various delayed operations, or to figure out the base of the log).

On that note, we flushed the vendored raticate libraries, replacing them with
tatami_r. We also updated the version of the tatami core library. We provide a
Rtatami.h header that defines all the macros consistently; developers are
expected to use this header rather than including the individual projects
manually, as the order of includes is important.