Try to automatically load and use beachmat.hdf5 for HDF5Array classes.
This should improve the seamlessness of using beachmat extensions like
beachmat.hdf5, provided it is installed on the user's machine. If not, we
just keep on going by falling back to the unknown matrix initializer.

Note that now we need to use a different matrix class for our unknown matrix
fallback tests, so we switch to a SolidRleArraySeed instead.